add pass on xml assertion
on atoum if I do something like this:
```
$this->string('mystring');
```
the score will be incremented and, if there is only that in the method
the result of the test will not be void.

When I use the extension to check if my serialize has worked (with without
checking if xml is valid against a schema, just to the if there is no error
(like adding directy a "&" in the xml without pass it by htmlspecialchars)),
the result of the test tells me that there is void test methods).

For example a method with just:
```
$this->xml("<?xml version=\"1.0\"?><root></root>");
```

Adding a call to pass in the setValue avoids that.